+++ sdk/identity/azure_identity/CHANGELOG.md
@@ -15,6 +15,7 @@
 ### Bugs Fixed
 
 - `ClientCertificateCredential::get_token()` returned an error when given multiple scopes
+- `ManagedIdentityCredential` didn't follow IMDS retry guidance
 
 ### Other Changes