+++ sdk/identity/azure_identity/CHANGELOG.md
@@ -1,5 +1,15 @@
 # Release History
 
+## 0.31.0 (Unreleased)
+
+### Features Added
+
+### Breaking Changes
+
+### Bugs Fixed
+
+### Other Changes
+
 ## 0.30.0 (2025-11-11)
 
 ### Features Added